More fixes. Refs #258
git-svn-id: https://svn.nexusformat.org/code/trunk@1648 ff5d1e40-2be0-497f-93bd-dc18237bd3c7

+++ src/nxdataset.c
@@ -33,7 +33,7 @@ static int getTypeSize(int typecode){
   }
 }
 /*-----------------------------------------------------------------------*/
-pNXDS createNXDataset(int rank, int typecode, int dim[]){
+pNXDS createNXDataset(int rank, int typecode, int64_t dim[]){
   pNXDS pNew = NULL;
   int i, length;
